[Surrey] Treat surreycpt user as if it were surreycc
The 'surreycpt' user was experience problems because it was
the only user that didn't have a matching entry in the body
table and the OPTION_SITE_NAME config.
This caused SQL problems with offline petitions, in that
the false value being returned from cobrand_admin_is_site_user
was being treated as a string in the SQL query. Adding
surreycpt to the OPTION_SITE_NAME resolved this, but still
didn't allow offline petitions to be created because there
was no body to attach them to.

This commit fixes the problem by hardcoding a special case
that treats the surreycpt as if it were the surreycc user
within cobrand_admin_is_site_user.