[RBWM] Fix bug whereby all petitions were updated.
No where was given on the update (misled by the ID being provided to the
subselect) and so on each iteration of the loop, every petition was
updated to have a cached_signers of 1 and an updated offline_signers
with something incorrect. At least only the correct signer rows were
deleted.