Daemons: provide debug, foreground and daemon modes
This allows the daemons to run in one of three modes: `--debug` and
`--foreground` both run in the foreground and log solely to STDERR;
`--debug` also sets a flag used in petemaild to trigger constant
database scans; the default is still to daemonize but now this just
calls `mySociety::SystemMisc::daemon()` once.

Separating out `--debug` from `--foreground` provides us with a
mode of running that works nicely with systemd which takes care of
process management and logging directly.